# frigappapi/src/main/java/com/frigapp/frigappapi/controller/api/AuthApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.frigapp.frigappapi.controller.api;

import com.frigapp.frigappapi.model.*;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.context.request.NativeWebRequest;

import java.util.Optional;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-26T22:36:13.511413300+01:00[Europe/Paris]", comments = "Generator version: 7.4.0")
@Validated
@Tag(name = "Auth", description = "Inscription, connexion et gestion du token JWT")
public interface AuthApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /auth/login : Connexion
     * Authentifie un utilisateur et retourne un token JWT.
     *
     * @param loginRequest  (required)
     * @return Connexion réussie (status code 200)
     *         or Données invalides (status code 400)
     *         or Identifiants incorrects (status code 401)
     */
    @Operation(
        operationId = "authLoginPost",
        summary = "Connexion",
        description = "Authentifie un utilisateur et retourne un token JWT.",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Connexion réussie", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuthLoginPost200Response.class))
            }),
            @ApiResponse(responseCode = "400", description = "Données invalides", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Identifiants incorrects", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/auth/login",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<AuthLoginPost200Response> authLoginPost(
        @Parameter(name = "LoginRequest", description = "", required = true) @Valid @RequestBody LoginRequest loginRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : { \"expiresIn\" : 3600, \"user\" : { \"firstName\" : \"Nathanaël\", \"id\" : \"550e8400-e29b-41d4-a716-446655440000\", \"email\" : \"nathanael@email.com\" }, \"token\" : \"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\" }, \"status\" : \"success\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /auth/refresh : Rafraîchir le token
     * Génère un nouveau token JWT à partir d&#39;un token valide.
     *
     * @return Token renouvelé (status code 200)
     *         or Token JWT absent ou invalide (status code 401)
     */
    @Operation(
        operationId = "authRefreshPost",
        summary = "Rafraîchir le token",
        description = "Génère un nouveau token JWT à partir d'un token valide.",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Token renouvelé", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuthLoginPost200Response.class))
            }),
            @ApiResponse(responseCode = "401", description = "Token JWT absent ou invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/auth/refresh",
        produces = { "application/json" }
    )
    
    default ResponseEntity<AuthLoginPost200Response> authRefreshPost(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : { \"expiresIn\" : 3600, \"user\" : { \"firstName\" : \"Nathanaël\", \"id\" : \"550e8400-e29b-41d4-a716-446655440000\", \"email\" : \"nathanael@email.com\" }, \"token\" : \"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\" }, \"status\" : \"success\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /auth/register : Créer un compte
     * Inscrit un nouvel utilisateur. Retourne le profil créé.
     *
     * @param registerRequest  (required)
     * @return Compte créé avec succès (status code 201)
     *         or Données invalides (status code 400)
     *         or Email déjà utilisé (status code 409)
     */
    @Operation(
        operationId = "authRegisterPost",
        summary = "Créer un compte",
        description = "Inscrit un nouvel utilisateur. Retourne le profil créé.",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Compte créé avec succès", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuthRegisterPost201Response.class))
            }),
            @ApiResponse(responseCode = "400", description = "Données invalides", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "409", description = "Email déjà utilisé", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/auth/register",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<AuthRegisterPost201Response> authRegisterPost(
        @Parameter(name = "RegisterRequest", description = "", required = true) @Valid @RequestBody RegisterRequest registerRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : { \"firstName\" : \"Nathanaël\", \"createdAt\" : \"2026-03-23T10:00:00Z\", \"id\" : \"550e8400-e29b-41d4-a716-446655440000\", \"email\" : \"nathanael@email.com\" }, \"status\" : \"success\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}