# frigappapi/src/main/resources/openapi/openapi_frigapp_v1.0.yaml
openapi: 3.0.3
info:
  title: Frig'App API
  description: |
    API REST de Frig'App — application mobile de gestion de stock de frigo collaboratif.
    
    ## Authentification
    Toutes les routes protégées nécessitent un token JWT passé dans le header `Authorization: Bearer <token>`.
    
    ## Temps réel
    Les modifications sur les items et les membres sont diffusées via WebSocket (STOMP over SockJS) sur `ws://api.frigapp.com/v1/ws`.
    
    ## Conventions
    - Toutes les réponses sont enveloppées dans `{ "status": "success"|"error", "data": ... }`
    - Les identifiants sont des UUID v4
    - Les timestamps sont en ISO 8601 UTC
  version: 1.0.0
  contact:
    name: Équipe Frig'App
servers:
  - url: https://api.frigapp.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Développement local (Spring Boot)
...
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── SCHEMAS ─────────────────────────────────────────────────────────────────
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
          example: success
        data:
          nullable: true
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
          example: error
        error:
          type: object
          properties:
            code:
              type: string
              example: FRIDGE_NOT_FOUND
            message:
              type: string
              example: Le frigo demandé n'existe pas.

    # ── USER ──────────────────────────────────────────────────────────────────
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        firstName:
          type: string
          example: Nathanaël
        email:
          type: string
          format: email
          example: nathanael@email.com
    UserFull:
      allOf:
        - $ref: '#/components/schemas/UserSummary'
        - type: object
          properties:
            createdAt:
              type: string
              format: date-time
              example: 2026-03-23T10:00:00Z

    # ── AUTH ──────────────────────────────────────────────────────────────────
    RegisterRequest:
      type: object
      required: [firstName, email, password]
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 50
          example: Nathanaël
        email:
          type: string
          format: email
          example: nathanael@email.com
        password:
          type: string
          minLength: 8
          example: motdepasse123
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: nathanael@email.com
        password:
          type: string
          example: motdepasse123
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        expiresIn:
          type: integer
          example: 3600
        user:
          $ref: '#/components/schemas/UserSummary'

    # ── FRIDGE ────────────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Token JWT absent ou invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            error:
              code: UNAUTHORIZED
              message: Token JWT absent ou invalide.
...
    BadRequest:
      description: Données invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

# ── PATHS ─────────────────────────────────────────────────────────────────────
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Créer un compte
      description: Inscrit un nouvel utilisateur. Retourne le profil créé.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Compte créé avec succès
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserFull'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email déjà utilisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: error
                error:
                  code: EMAIL_ALREADY_EXISTS
                  message: Un compte existe déjà avec cet email.
  /auth/login:
    post:
      tags: [Auth]
      summary: Connexion
      description: Authentifie un utilisateur et retourne un token JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Identifiants incorrects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: error
                error:
                  code: INVALID_CREDENTIALS
                  message: Email ou mot de passe incorrect.
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rafraîchir le token
      description: Génère un nouveau token JWT à partir d'un token valide.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token renouvelé
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── FRIDGES ───────────────────────────────────────────────────────────────
...